moving BasicUserWithFriendStatus out of LibraryView.scala
